Greatly improved line-to-word segmentation. Updated README.md.
Renamed lineAnalysis.py to wordAnalysis.py and very much improved the file: Creating a much more robust method for segmenting line images into series' of word images. Updated the README.md to reflect these changes, as well as the changes made previously in GMM line detection. Removed dilation step, as it was unnecessary with the improved methods, and sometimes harmful to the analysis.